fixed class and tests

+++ src/main/java/RewardValue.java
@@ -2,10 +2,10 @@ import java.util.Scanner;
 
 public class RewardValue {
     private double cashValue;
-    private double milesValue;
+    private int milesValue;
     public RewardValue(double cashValue){
         this.cashValue = cashValue;
-        this.milesValue = cashValue / 0.0035;
+        this.milesValue = (int) (cashValue / 0.0035);
     }
     public RewardValue(int milesValue){
         this.milesValue = milesValue;
